Change Rule from an interface to a struct
