+++ main.go
@@ -8,6 +8,8 @@ import (
 	"alphanonce.com/exchangesimulator/internal/log"
 	"alphanonce.com/exchangesimulator/internal/log/config"
 	"alphanonce.com/exchangesimulator/internal/rule"
+	"alphanonce.com/exchangesimulator/internal/rule/request_matcher"
+	"alphanonce.com/exchangesimulator/internal/rule/responder"
 	"alphanonce.com/exchangesimulator/internal/server"
 	"alphanonce.com/exchangesimulator/internal/simulator"
 )
@@ -26,18 +28,14 @@ func init() {
 
 func main() {
 	rules := []rule.Rule{
-		rule.NewStaticResponseRule(
-			"GET",
-			"/api/v4/public/platform/status",
-			`{"status":"1"}`,
-			time.Second,
-		),
-		rule.NewStaticResponseRule(
-			"GET",
-			"/api/v4/public/ping",
-			`["pong"]`,
-			time.Second,
-		),
+		{
+			RequestMatcher: request_matcher.NewRequestPredicate("GET", "/api/v4/public/platform/status"),
+			Responder:      responder.NewResponseFromString(`{"status":"1"}`, time.Second),
+		},
+		{
+			RequestMatcher: request_matcher.NewRequestPredicate("GET", "/api/v4/public/ping"),
+			Responder:      responder.NewResponseFromString(`["pong"]`, time.Second),
+		},
 	}
 	s := simulator.NewSimulator(rules)
 	sv := server.NewFasthttpServer(s)
